# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Timeless Love API
  description: Backend API for Timeless Love family social platform
  version: 1.0.0
  contact:
    name: API Support
    email: support@timelesslove.com
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.timelesslove.com
    description: Production server
...
paths:
...
  /api/v1/invites:
    post:
      tags:
        - invites
      summary: Create Invitation
      description: Create a new invitation (Adult only)
      operationId: createInvite
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInviteRequest'
      responses:
        '201':
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/invites/{token}:
    get:
      tags:
        - invites
      summary: Validate Invitation
      description: Validate an invitation token
      operationId: validateInvite
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
          description: Invitation token
      responses:
        '200':
          description: Invitation validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateInviteResponse'
  /api/v1/invites/{token}/accept:
    post:
      tags:
        - invites
      summary: Accept Invitation
      description: Accept an invitation and create user account
      operationId: acceptInvite
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
          description: Invitation token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptInviteRequest'
      responses:
        '200':
          description: Invitation accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateInviteRequest:
      type: object
      required:
        - email
        - role
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          enum: [adult, teen, child, grandparent]
    InviteResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
        token:
          type: string
        invite_link:
          type: string
        expires_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, accepted, expired, revoked]
        created_at:
          type: string
          format: date-time
    ValidateInviteResponse:
      type: object
      properties:
        valid:
          type: boolean
        invite:
          $ref: '#/components/schemas/InviteResponse'
        message:
          type: string
    AcceptInviteRequest:
      type: object
      required:
        - email
        - password
        - display_name
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        display_name:
          type: string
          minLength: 1
...
    Error:
      type: object
      properties:
        detail:
          type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...